backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "An ICICI Breeze Api wrapper for Python"
authors = [
    { name = "Mohammad Naveed Khan", email = "naveedkhan13041998@gmail.com" },
]
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "amqp>=5.3.1",
    "arrow>=1.3.0",
    "asgiref>=3.8.1",
    "async-timeout>=5.0.1",
    "attrs>=25.3.0",
    "autobahn>=24.4.2",
    "automat>=25.4.16",
    "bidict>=0.23.1",
    "billiard>=4.2.1",
    "binaryornot>=0.4.4",
    "breeze-connect>=1.0.63",
    "celery>=5.5.3",
    "certifi>=2025.4.26",
    "cffi>=1.17.1",
    "channels>=4.2.2",
    "chardet>=5.2.0",
    "charset-normalizer>=3.4.2",
    "click>=8.2.1",
    "click-didyoumean>=0.3.1",
    "click-plugins>=1.1.1",
    "click-repl>=0.3.0",
    "colorama>=0.4.6",
    "constantly>=23.10.4",
    "cookiecutter>=2.6.0",
    "cryptography>=45.0.4",
    "cython>=3.1.2",
    "daphne>=4.2.0",
    "django>=5.2.3",
    "django-browser-reload>=1.18.0",
    "django-cors-headers>=4.7.0",
    "django-filter>=25.1",
    "django-redis>=5.4.0",
    "django-tailwind>=4.0.1",
    "djangorestframework>=3.16.0",
    "djangorestframework-simplejwt>=5.5.0",
    "flower>=2.0.1",
    "gunicorn>=23.0.0",
    "hyperlink>=21.0.0",
    "idna>=3.10",
    "incremental>=24.7.2",
    "jinja2>=3.1.6",
    "kombu>=5.5.4",
    "markdown-it-py>=3.0.0",
    "markupsafe>=3.0.2",
    "mdurl>=0.1.2",
    "numpy>=2.3.0",
    "pandas>=2.3.0",
    "pillow>=11.2.1",
    "prompt-toolkit>=3.0.51",
    "psycopg2>=2.9.10",
    "pyasn1>=0.6.1",
    "pyasn1-modules>=0.4.2",
    "pycparser>=2.22",
    "pygments>=2.19.1",
    "pyjwt>=2.9.0",
    "pyopenssl>=25.1.0",
    "python-dateutil>=2.9.0.post0",
    "python-engineio>=4.12.2",
    "python-slugify>=8.0.4",
    "python-socketio>=5.13.0",
    "pytz>=2025.2",
    "pyyaml>=6.0.2",
    "redis>=6.2.0",
    "requests>=2.32.4",
    "rich>=14.0.0",
    "service-identity>=24.2.0",
    "six>=1.17.0",
    "sqlparse>=0.5.3",
    "text-unidecode>=1.3",
    "twisted>=25.5.0",
    "txaio>=23.1.1",
    "types-python-dateutil>=2.9.0.20250516",
    "typing-extensions>=4.14.0",
    "tzdata>=2025.2",
    "urllib3>=2.4.0",
    "vine>=5.1.0",
    "watchdog>=6.0.0",
    "wcwidth>=0.2.13",
    "websocket-client>=1.8.0",
    "zope-interface>=7.2",
]

[project.optional-dependencies]
dev = ["ruff>=0.1.0", "black>=23.0.0"]

[tool.black]
line-length = 88
target-version = ['py313']
include = '\.pyi?$'
extend-exclude = '''
/(
  # directories
  \.eggs
  | \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | build
  | dist
  | migrations
)/
'''

[tool.ruff]
target-version = "py313"
line-length = 88
select = [
    "E",      # pycodestyle errors
    "W",      # pycodestyle warnings
    "F",      # pyflakes
    "I",      # isort
    "B",      # flake8-bugbear
    "C4",     # flake8-comprehensions
    "UP",     # pyupgrade
    "ARG001", # unused-function-args
    "SIM",    # flake8-simplify
    "TC",     # flake8-type-checking
    "TID",    # flake8-tidy-imports
    "Q",      # flake8-quotes
    "UP",     # pyupgrade
    "PTH",    # flake8-use-pathlib
]
ignore = [
    "E501", # line too long, handled by black
    "B008", # do not perform function calls in argument defaults
    "C901", # too complex
    "W191", # indentation contains tabs
]
unfixable = [
    "F841", # remove unused variables
]
exclude = [
    ".bzr",
    ".direnv",
    ".eggs",
    ".git",
    ".git-rewrite",
    ".hg",
    ".mypy_cache",
    ".nox",
    ".pants.d",
    ".pytype",
    ".ruff_cache",
    ".svn",
    ".tox",
    ".venv",
    "__pypackages__",
    "_build",
    "buck-out",
    "build",
    "dist",
    "node_modules",
    "venv",
    "migrations",
]
dummy-variable-rgx = "^(_+|(_+[a-zA-Z0-9_]*[a-zA-Z0-9]+?))$"

[tool.ruff.mccabe]
max-complexity = 10

[tool.ruff.per-file-ignores]
"__init__.py" = ["F401"]
"**/tests/**/*" = ["S101", "ARG", "FBT"]
"**/migrations/**/*" = ["ALL"]

[tool.ruff.isort]
known-first-party = ["backend"]
force-sort-within-sections = true
split-on-trailing-comma = true
